make scheduler_perf usable from other repositories
